refactor map screens with shared geojson helpers and effect events
